As a step towards frictionless end user experiences when using ShouldThrow, add non-implemented overloads of ShouldThrow along with incomplete tests that do completely hit all intended coverage combinations.
As of this commit the solution builds, but the tests are not expected to pass. Many of the new tests' "Contradiction" assertions include an empty string as the expected message, so that the many similar cases can be visually scanned for completeness before the real expected strings get filled in.