Revert "Simplify dynamic method invocation, and benefit from the performance optimizations provided by `dynamic`."
This reverts commit ef313bd63b29630745b2219f228a67145872d3a5.

This use of `dynamic` is unsound. For instance, if the return type of the Func<T> in question happens to be a custom type like an empty struct, the success or failure of the dynamic invocation can vary, with a very misleading exception message, based on whether the struct is defined publicly vs for instance nesting within the test class where it is used. It's as if visibility of the types involved affects what happens, so whatever the actual explanation is, using `dynamic` in this way simply is not the same as calling DynamicInvoke.